organize snarls by chr earlier, rename stuff
- rename Snarl_data_t variables
  - snarl -> net
  - snarl_paths -> path
  - snarl_ids -> ids
- rename ref_chr to ref_chrs
- remove find_snarl_id function, not needed anymore
- use Snarl_data_t organized by chrom earlier
  - tuple removed!
  - make sure it's written by chromosome to avoid potential bug (see below)
- filter snarls not on chr of interest earlier

I think there was a small risk of the snarl file being written wrong
if multithreading was used. A few record of the current chromosome
could end up after the new one started which would mean breaking the
assumption that snarls are sorted by chromosome (when reading it). By
listing them by chromosome from the start, not only do we avoid
carrying a 'chr' info with each snarl, we make sure they are really
written together.